# backend/app/api/v1/endpoints/auth.py
from typing import Any
from datetime import timedelta
from fastapi import APIRouter, Depends, HTTPException, status
from fastapi.security import OAuth2PasswordRequestForm
from sqlalchemy.ext.asyncio import AsyncSession

from app.core import security
from app.core.config import settings
from app.core.database import get_db
from app.infrastructure.db.repositories.user_repository import UserRepository
from app.models.user import User
from pydantic import BaseModel, EmailStr

router = APIRouter()

class UserCreate(BaseModel):
    email: EmailStr
    password: str

class Token(BaseModel):
    access_token: str
    token_type: str

@router.post("/login", response_model=Token)
async def login_access_token(
    db: AsyncSession = Depends(get_db), form_data: OAuth2PasswordRequestForm = Depends()
) -> Any:
    """
    用户登录接口 (获取 Access Token)
    1. 验证邮箱和密码
    2. 生成 JWT Token
    """
    # 查询用户
    user = await UserRepository(db).get_by_email(form_data.username)
    
    # 验证密码
    if not user or not security.verify_password(form_data.password, user.hashed_password):
        raise HTTPException(
            status_code=status.HTTP_400_BAD_REQUEST,
            detail="Incorrect email or password",
        )
    
    # 生成 Token
    access_token_expires = timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
    access_token = security.create_access_token(
        user.id, expires_delta=access_token_expires
    )
    return {"access_token": access_token, "token_type": "bearer"}

@router.post("/register", response_model=Token)
async def register_user(
    user_in: UserCreate, db: AsyncSession = Depends(get_db)
) -> Any:
    """
    用户注册接口
    1. 检查邮箱是否已存在
    2. 创建新用户并加密密码
    3. 自动登录 (返回 Token)
    """
    # 检查邮箱是否存在
    repo = UserRepository(db)
    existing_user = await repo.get_by_email(user_in.email)
    
    if existing_user:
        raise HTTPException(
            status_code=400,
            detail="The user with this email already exists in the system.",
        )
    
    # 创建用户
    user = User(
        email=user_in.email,
        hashed_password=security.get_password_hash(user_in.password)
    )
    user = await repo.create(user)
    
    # 注册成功后直接签发 Token，实现自动登录
    access_token_expires = timedelta(minutes=settings.ACCESS_TOKEN_EXPIRE_MINUTES)
    access_token = security.create_access_token(
        user.id, expires_delta=access_token_expires
    )
    return {"access_token": access_token, "token_type": "bearer"}