# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Smart Investment Advisor",
    "description": "AI 智能投资助手后端 API，集成多源数据与 LLM 分析能力",
    "version": "1.0.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Login Access Token",
        "description": "用户登录接口 (获取 Access Token)\n1. 验证邮箱和密码\n2. 生成 JWT Token",
        "operationId": "login_access_token_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_login_access_token_api_auth_login_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Register User",
        "description": "用户注册接口\n1. 检查邮箱是否已存在\n2. 创建新用户并加密密码\n3. 自动登录 (返回 Token)",
        "operationId": "register_user_api_auth_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/portfolio/": {
      "get": {
        "tags": [
          "portfolio"
        ],
        "summary": "Get Portfolio",
        "description": "获取当前用户的投资组合列表 (Fetch User Portfolio)\n\n逻辑 (Logic)：\n1. 联表查询：Portfolio (持仓) -> MarketDataCache (行情缓存) -> Stock (基础资料)\n2. 若 refresh=True，则同步/异步触发全量实时刷新\n3. 循环计算盈亏比例、市值、未实现盈亏等前端展示字段",
        "operationId": "get_portfolio_api_portfolio__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "refresh",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Refresh"
            }
          },
          {
            "name": "price_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Price Only"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PortfolioItem"
                  },
                  "title": "Response Get Portfolio Api Portfolio  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "portfolio"
        ],
        "summary": "Add Portfolio Item",
        "description": "添加或更新自选股/持仓",
        "operationId": "add_portfolio_item_api_portfolio__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/portfolio/{ticker}": {
      "delete": {
        "tags": [
          "portfolio"
        ],
        "summary": "Delete Portfolio Item",
        "description": "从自选股中删除",
        "operationId": "delete_portfolio_item_api_portfolio__ticker__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/analysis/portfolio": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Portfolio Analysis",
        "description": "获取最近的一次全量持仓健康分析 (非实时生成)",
        "operationId": "get_portfolio_analysis_api_analysis_portfolio_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioAnalysisResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Analyze Portfolio",
        "description": "全量持仓健康分析接口 (Portfolio Health Check)",
        "operationId": "analyze_portfolio_api_analysis_portfolio_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioAnalysisResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/analysis/{ticker}": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Analyze Stock",
        "description": "股票分析接口 (核心业务逻辑)\n1. 检查用户权限/SaaS限制 (免费用户每日限额)\n2. 获取实时行情数据 (价格、技术指标)\n3. 获取相关新闻 (上下文)\n4. 获取用户持仓信息 (成本、盈亏)\n5. 调用 AI 模型生成投资建议",
        "operationId": "analyze_stock_api_analysis__ticker__post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Force"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Latest Analysis",
        "description": "获取单支股票最新的分析记录（不触发新 AI 分析，仅查库）",
        "operationId": "get_latest_analysis_api_analysis__ticker__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/analysis/{ticker}/history": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Get Analysis History",
        "description": "获取股票历史分析记录列表 (History of AI Analysis)",
        "operationId": "get_analysis_history_api_analysis__ticker__history_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalysisResponse"
                  },
                  "title": "Response Get Analysis History Api Analysis  Ticker  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_login_access_token_api_auth_login_post": {
        "properties": {
          "grant_type": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^password$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grant Type"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "password": {
            "type": "string",
            "format": "password",
            "title": "Password"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": ""
          },
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          },
          "client_secret": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "format": "password",
            "title": "Client Secret"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "Body_login_access_token_api_auth_login_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PortfolioAnalysisResponse": {
        "properties": {
          "health_score": {
            "type": "integer",
            "title": "Health Score"
          },
          "risk_level": {
            "type": "string",
            "title": "Risk Level"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "diversification_analysis": {
            "type": "string",
            "title": "Diversification Analysis"
          },
          "strategic_advice": {
            "type": "string",
            "title": "Strategic Advice"
          },
          "top_risks": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Top Risks"
          },
          "top_opportunities": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Top Opportunities"
          },
          "detailed_report": {
            "type": "string",
            "title": "Detailed Report"
          },
          "model_used": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Used"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "health_score",
          "risk_level",
          "summary",
          "diversification_analysis",
          "strategic_advice",
          "top_risks",
          "top_opportunities",
          "detailed_report",
          "created_at"
        ],
        "title": "PortfolioAnalysisResponse"
      },
      "PortfolioCreate": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "avg_cost": {
            "type": "number",
            "title": "Avg Cost"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "quantity",
          "avg_cost"
        ],
        "title": "PortfolioCreate"
      },
      "PortfolioItem": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "quantity": {
            "type": "number",
            "title": "Quantity"
          },
          "avg_cost": {
            "type": "number",
            "title": "Avg Cost"
          },
          "current_price": {
            "type": "number",
            "title": "Current Price",
            "default": 0.0
          },
          "market_value": {
            "type": "number",
            "title": "Market Value",
            "default": 0.0
          },
          "unrealized_pl": {
            "type": "number",
            "title": "Unrealized Pl",
            "default": 0.0
          },
          "pl_percent": {
            "type": "number",
            "title": "Pl Percent",
            "default": 0.0
          },
          "last_updated": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Updated"
          },
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "industry": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Industry"
          },
          "market_cap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Cap"
          },
          "pe_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pe Ratio"
          },
          "forward_pe": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forward Pe"
          },
          "eps": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eps"
          },
          "dividend_yield": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dividend Yield"
          },
          "beta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Beta"
          },
          "fifty_two_week_high": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fifty Two Week High"
          },
          "fifty_two_week_low": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fifty Two Week Low"
          },
          "pe_percentile": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pe Percentile"
          },
          "pb_percentile": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pb Percentile"
          },
          "net_inflow": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Inflow"
          },
          "rsi_14": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rsi 14"
          },
          "ma_20": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ma 20"
          },
          "ma_50": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ma 50"
          },
          "ma_200": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ma 200"
          },
          "macd_val": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Macd Val"
          },
          "macd_signal": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Macd Signal"
          },
          "macd_hist": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Macd Hist"
          },
          "macd_hist_slope": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Macd Hist Slope"
          },
          "macd_cross": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Macd Cross"
          },
          "macd_is_new_cross": {
            "type": "boolean",
            "title": "Macd Is New Cross",
            "default": false
          },
          "bb_upper": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bb Upper"
          },
          "bb_middle": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bb Middle"
          },
          "bb_lower": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bb Lower"
          },
          "atr_14": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Atr 14"
          },
          "k_line": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "K Line"
          },
          "d_line": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "D Line"
          },
          "j_line": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "J Line"
          },
          "volume_ma_20": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume Ma 20"
          },
          "volume_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume Ratio"
          },
          "adx_14": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adx 14"
          },
          "pivot_point": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pivot Point"
          },
          "resistance_1": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resistance 1"
          },
          "resistance_2": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resistance 2"
          },
          "support_1": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Support 1"
          },
          "support_2": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Support 2"
          },
          "risk_reward_ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Reward Ratio"
          },
          "change_percent": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Percent",
            "default": 0.0
          },
          "market_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Status"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "quantity",
          "avg_cost"
        ],
        "title": "PortfolioItem"
      },
...
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "token_type"
        ],
        "title": "Token"
      },
...
      "UserCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UserCreate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}